// src/lib/enums.ts
export enum Criterion {
  HighRemuneration = 'high-remuneration',
  Membership = 'membership',
  OriginalContributions = 'original-contributions',
  CriticalRole = 'critical-role',
  Awards = 'awards',
  Press = 'press',
  Judging = 'judging',
  ScholarlyArticles = 'scholarly-articles',
}

export enum Status {
  InProgress = 'in-progress',
  Complete = 'complete',
  NotStarted = 'not-completed',
  ReadyForReview = 'ready-for-review',
}